huge refactor before merging back into master
