+++ infrastructure/ansible/supervisor/templates/server.conf
@@ -1,5 +1,5 @@
 [program:server]
-command=gunicorn app:app -c /home/ansible/app/gunicorn.conf.py
+command=gunicorn linkalytics:app -c /home/ansible/linkalytics/gunicorn.conf.py
 process_name=%(program_name)s
 
 directory=/home/ansible
