+++ infrastructure/ansible/supervisor/templates/worker.conf
@@ -1,5 +1,5 @@
 [program:queue_worker]
-command=/usr/bin/python3 -m src {{ queues }}
+command=/usr/bin/python3 -m linkalytics {{ queues }}
 process_name=%(program_name)s
 
 directory=/home/ansible
