+++ linkalytics/__init__.py
@@ -8,4 +8,6 @@ from .__version__ import __build__
 __title__     = 'linkalytics'
 __license__   = 'Apache Software License Version 2.0'
 
-from . server import app
+from .app import app
+
+
