+++ linkalytics/coincidence/__init__.py
@@ -1,0 +1,3 @@
+from . coincidence import run
+from . coincidence import specific_term
+from . coincidence import unique_features
