+++ linkalytics/lsh/__init__.py
@@ -1,1 +1,2 @@
 from . lsh import run
+from . lsh import lsh
