+++ linkalytics/ngrams/ngrams.py
@@ -1,11 +1,11 @@
 import functools
 
-from .. tdm import TermDocumentMatrix
-from .. tdm import n_grams
-from .. tdm import get_results
+from .. search import get_results
+from .  entropy import TermDocumentMatrix, ngrams
+
 
 def run(node):
-    tokenizer = functools.partial(n_grams, numbers=True, normalize=True)
+    tokenizer = functools.partial(ngrams, numbers=True, normalize=True)
     tdm       = TermDocumentMatrix(cutoff=1, tokenizer=tokenizer)
     results   = get_results(node.get('text', None), node.get('size', 1000), True)
