+++ linkalytics/run/__init__.py
@@ -1,0 +1,6 @@
+class Arguments:
+
+    def __init__(self, query, size):
+        self.query = [query]
+        self.size  = [size]
+
