+++ linkalytics/worker.py
@@ -5,7 +5,7 @@ import logging
 
 from multiprocessing import cpu_count
 from .environment import cfg
-from .task_mux import TaskMux
+from .tasks import TaskMux
 from . import instagrammer
 from . import phonenumber
 from . import twitter
