+++ manage.py
@@ -5,7 +5,7 @@ from flask.ext.script import Manager, Server
 try:
     from app import app
 except ImportError:
-    from src import app
+    from linkalytics import app
 
 
 server  = Server(host="0.0.0.0", port=8080)
