--- src/coincidence/__init__.py
@@ -1,1 +1,0 @@
-from . coincidence import run
