--- src/ngrams/__init__.py
@@ -1,1 +1,0 @@
-from . ngrams import run
