--- src/test_server.py
@@ -1,10 +1,0 @@
-# import nose
-# from utils import with_test
-# from server import doc_to_group, query_docs, look_up, post_new
-
-
-# @with_test(doc_to_group, 'server')
-# def data():
-#     yield ({"text": "14623914"}, [
-#         {"_id": "14072230210", "_type": "group"}
-#         ])
